refactor: rename cmake_settings -> skbuild_settings
