+++ tests/test_skbuild_settings.py
@@ -1,9 +1,9 @@
 import textwrap
 
-from scikit_build_core.settings.cmake_settings import read_settings
+from scikit_build_core.settings.skbuild_settings import read_settings
 
 
-def test_cmake_settings_default(tmp_path):
+def test_skbuild_settings_default(tmp_path):
     pyproject_toml = tmp_path / "pyproject.toml"
     pyproject_toml.write_text("", encoding="utf-8")
 
@@ -15,7 +15,7 @@ def test_cmake_settings_default(tmp_path):
     assert settings.ninja.min_version == "0.0"
 
 
-def test_cmake_settings_envvar(tmp_path, monkeypatch):
+def test_skbuild_settings_envvar(tmp_path, monkeypatch):
     monkeypatch.setenv("SKBUILD_CMAKE_MIN_VERSION", "3.16")
     monkeypatch.setenv("SKBUILD_NINJA_MIN_VERSION", "1.1")
 
@@ -30,7 +30,7 @@ def test_cmake_settings_envvar(tmp_path, monkeypatch):
     assert settings.ninja.min_version == "1.1"
 
 
-def test_cmake_settings_config_settings(tmp_path):
+def test_skbuild_settings_config_settings(tmp_path):
     pyproject_toml = tmp_path / "pyproject.toml"
     pyproject_toml.write_text("", encoding="utf-8")
 
@@ -45,7 +45,7 @@ def test_cmake_settings_config_settings(tmp_path):
     assert settings.ninja.min_version == "1.2"
 
 
-def test_cmake_settings_pyproject_toml(tmp_path):
+def test_skbuild_settings_pyproject_toml(tmp_path):
     pyproject_toml = tmp_path / "pyproject.toml"
     pyproject_toml.write_text(
         textwrap.dedent(
